add UsersServiceTest

+++ src/main/java/de/telran/shoponline/repository/OrderItemsRepository.java
@@ -3,6 +3,6 @@ package de.telran.shoponline.repository;
 import de.telran.shoponline.entity.OrderItems;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemsRepository extends JpaRepository<OrderItems, Integer> {
+public interface OrderItemsRepository extends JpaRepository<OrderItems, Long> {
 
 }
